WB-2005 content transformer client parametrized from infra config
